create initial dashboard files
