+++ dashboard/requirements.txt
@@ -3,3 +3,4 @@ pandas
 psycopg2-binary
 numpy
 plotly
+dash-bootstrap-components==1.6.0
